Extracted some of the test methods to a dedicated ClassPathParserTests class
